refactor: 코드 리팩토링 (커스텀 훅 생성)
